# perf: update benchmarks for 1.x