The expectation with the 0.x to 1.x migration is that changes to the API would not affect performance. With the understanding that noise has not been entirely controlled with runs, the comparisons below show that v0.60.1 is between 1.0 and 1.15 times slower than v1.0.0.rc3.

## Results

| Benchmark                                | 1.0.0rc3 | 0.60.1                |
|------------------------------------------|----------|-----------------------|
| encoder-one-trace                        | 324 us   | 373 us: 1.15x slower  |
| encoder-many-traces                      | 37.0 ms  | 41.2 ms: 1.11x slower |
| encoder-one-tag                          | 456 us   | 500 us: 1.10x slower  |
| encoder-one-metric                       | 461 us   | 496 us: 1.08x slower  |
| encoder-many-metrics                     | 1.92 ms  | 1.97 ms: 1.02x slower |
| encoder-one-trace-with-dd-origin         | 730 us   | 778 us: 1.07x slower  |
| encoder-one-trace-with-dd-origin-no-tags | 427 us   | 477 us: 1.12x slower  |
| encoder-many-traces-with-dd-origin       | 84.5 ms  | 88.8 ms: 1.05x slower |
| span-start                               | 2.62 ms  | 2.94 ms: 1.12x slower |
| span-add-tags                            | 63.9 ms  | 64.8 ms: 1.01x slower |
| span-start-finish                        | 4.09 ms  | 4.43 ms: 1.08x slower |
| tracer-small                             | 120 us   | 124 us: 1.04x slower  |
| tracer-medium                            | 1.05 ms  | 1.09 ms: 1.04x slower |
| tracer-large                             | 10.7 ms  | 11.2 ms: 1.04x slower |
| flasksimple-baseline                     | 1.56 ms  | 1.60 ms: 1.03x slower |
| flasksimple-profiler                     | 1.61 ms  | 1.60 ms: 1.01x faster |
| flasksimple-tracer-and-profiler          | 1.55 ms  | 1.60 ms: 1.04x slower |